move part of lib files to components
