+++ app-ui/src/pages/home/Home.tsx
@@ -2,7 +2,7 @@ import React, { useEffect } from "react";
 import { useUserInfo } from "../../utils/UserInfo";
 import { BsGlobe, BsDatabase, BsJustify } from "react-icons/bs";
 import { Link, useNavigate } from "react-router-dom";
-import { Modal } from "../../lib/Modal";
+import { Modal } from "../../components/Modal";
 import { Time } from "../../components/Time";
 import { useDatasetList } from "../../service/DatasetOperations";
 import { UploadDatasetModalContent } from "./NewDataset";
