+++ app-ui/src/pages/snippets/Snippets.tsx
@@ -1,5 +1,5 @@
 import React, { useState, useEffect, useCallback } from "react";
-import { EntityList } from "../../lib/EntityList";
+import { EntityList } from "../../components/EntityList";
 import { useUserInfo } from "../../utils/UserInfo";
 import { BsUpload } from "react-icons/bs";
 import { Link, useNavigate } from "react-router-dom";
